分数检测

+++ emc-dao/src/main/java/com/huak/health/model/HealthScoreRecord.java
@@ -9,13 +9,13 @@ public class HealthScoreRecord implements Serializable{
 
     private String orgId;
 
-    private Integer score;
+    private Double score;
 
     private String userid;
 
     private String createTime;
 
-    public HealthScoreRecord(String id, String orgId, Integer score, String userid, String createTime) {
+    public HealthScoreRecord(String id, String orgId, Double score, String userid, String createTime) {
         this.id = id;
         this.orgId = orgId;
         this.score = score;
@@ -43,11 +43,11 @@ public class HealthScoreRecord implements Serializable{
         this.orgId = orgId == null ? null : orgId.trim();
     }
 
-    public Integer getScore() {
+    public Double getScore() {
         return score;
     }
 
-    public void setScore(Integer score) {
+    public void setScore(Double score) {
         this.score = score;
     }
